SPlit out the wrokshop/level selection into a different module then the main UI, Also made the selected level details its proper own class
